+++ changelog.md
@@ -9,6 +9,7 @@ A project that using SmartFusion2 to control a gyroscope MPU6050 on the GY521 mo
 -----------------------------------
 ## Unreleased v1.#.#(2023.07.##) @master
 ### New
+- create `num_packet` in `SysTick_Handler()`
 ### Changed
 ### Removed
 ### Deprecated
@@ -18,9 +19,10 @@ A project that using SmartFusion2 to control a gyroscope MPU6050 on the GY521 mo
 ### Reference
 
 -----------------------------------
-## Unreleased v1.0.0(2023.07.05) @master
+## BT-Gyro v1.0.0(2023.07.05) @master
 ### New
 - release v0.7.0
+
 
 -----------------------------------
 ## BT-Gyro v0.7.0(2023.07.05) @master