+++ changelog.md
@@ -10,6 +10,7 @@ A project that using SmartFusion2 to control a gyroscope MPU6050 on the GY521 mo
 ## Unreleased v1.#.#(2023.07.##) @master
 ### New
 1. create `num_packet` in `SysTick_Handler()`
+2. add `Num_packet` into packet
 ### Changed
 1. move `num_packet` from `SysTick_Handler()` to `MPU6050_t`
 2. rename `Sensitivity` in `Mpu6050_t` to `Resolution`